2-Basics Of LangChain/requirements.txt
ipykernel
pypdf
bs4
arxiv
pymupdf
wikipedia
chromadb
sentence_transformers
faiss-cpu
streamlit
langchain>=0.3.60
langchain-core>=0.3.60
langsmith==0.0.99
langchain-community==0.0.21
langchain-openai==0.0.8
langchain-groq>=0.3.0
langchain-ollama>=0.3.3
langchain-huggingface==0.2.0
langchain-chroma==0.2.4
langchain-text-splitters==0.0.1
fastapi>=0.105
pydantic==1.10.13
uvicorn
python-dotenv
sse-starlette==1.6.5